Fix API call Bug & modify background style
